requirements.txt
# Core dependencies
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0

# Audio processing
librosa>=0.8.1
soundfile>=0.10.3
sounddevice>=0.4.4

# Configuration and utilities
pyyaml>=5.4.1
tqdm>=4.62.0

# Optional: for advanced audio processing
# pesq>=0.0.3
# pystoi>=0.3.3

# Development dependencies (uncomment for development)
# pytest>=6.2.0
# black>=21.0.0
# flake8>=3.9.0
# mypy>=0.910